Tons of new stuff
EnemyManager implemented
- Enemies are now bound to room and are deactivated upon exit.
- Enemies will reset to their start position upon entering the room
TODO: EvilEye start pos can get bugged

Map boundaries are now updated on OnTriggerStay2D()

Player:
- Player now has a proper hit state
- Player now uses OnHitEffect for transparency effect

Parallax Background:
- Added 3 background materials
- Background now scroll along the camera in different speeds.
TODO: Needs further testing, also jitters A LOT for some reason.

+++ Assets/Scripts/CameraFollow2D.cs
@@ -7,8 +7,8 @@ public class CameraFollow2D : MonoBehaviour {
 
 	public Transform target;
     public bool hideCursor;
-    [Range(1, 10)] public int followSpeed;
-    [Range(1, 10)] public int transitionSpeed;
+    [Range(1, 10)] public float followSpeed;
+    [Range(1, 10)] public float transitionSpeed;
 
     Vector2 topLeft, bottomRight;
     bool transitioning = false;
@@ -25,7 +25,6 @@ public class CameraFollow2D : MonoBehaviour {
 
 		if (!transitioning) {
 			transform.position = Vector3.MoveTowards (transform.position, new Vector3 (posX, posY, transform.position.z), 0.64f * followSpeed);
-			//transform.position = new Vector3 (posX, posY, transform.position.z);
 		} else {
 			RoomTransition ();
 		}
